handling undefined label at the annotation statistics
