+++ tests/unit/test_ops.py
@@ -8,7 +8,6 @@ import pytest
 from datumaro.components.annotation import (
     AnnotationType,
     Bbox,
-    Caption,
     Ellipse,
     Label,
     LabelCategories,
@@ -25,7 +24,7 @@ from datumaro.components.media import Image, MultiframeImage, PointCloud
 from datumaro.components.merge.exact_merge import ExactMerge
 from datumaro.components.merge.intersect_merge import IntersectMerge
 from datumaro.components.merge.union_merge import UnionMerge
-from datumaro.components.operations import compute_ann_statistics, find_unique_images, mean_std
+from datumaro.components.operations import find_unique_images, mean_std
 from datumaro.errors import (
     ConflictingCategoriesError,
     FailedAttrVotingError,
@@ -66,277 +65,6 @@ class TestOperations(TestCase):
             assert np.allclose(em, am, atol=0.6)
         for estd, astd in zip(expected_std, actual_std):
             assert np.allclose(estd, astd, atol=0.1)
-
-    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_stats(self):
-        dataset = Dataset.from_iterable(
-            [
-                DatasetItem(
-                    id=1,
-                    media=Image.from_numpy(data=np.ones((5, 5, 3))),
-                    annotations=[
-                        Caption("hello"),
-                        Caption("world"),
-                        Label(
-                            2,
-                            attributes={
-                                "x": 1,
-                                "y": "2",
-                            },
-                        ),
-                        Bbox(
-                            1,
-                            2,
-                            2,
-                            2,
-                            label=2,
-                            attributes={
-                                "score": 0.5,
-                            },
-                        ),
-                        Bbox(
-                            5,
-                            6,
-                            2,
-                            2,
-                            attributes={
-                                "x": 1,
-                                "y": "3",
-                                "occluded": True,
-                            },
-                        ),
-                        Points([1, 2, 2, 0, 1, 1], label=0),
-                        Mask(
-                            label=3,
-                            image=np.array(
-                                [
-                                    [0, 0, 1, 1, 1],
-                                    [0, 0, 1, 1, 1],
-                                    [0, 0, 1, 1, 1],
-                                    [0, 0, 0, 0, 0],
-                                    [0, 0, 0, 0, 0],
-                                ]
-                            ),
-                        ),
-                    ],
-                ),
-                DatasetItem(
-                    id=2,
-                    media=Image.from_numpy(data=np.ones((2, 4, 3))),
-                    annotations=[
-                        Label(
-                            2,
-                            attributes={
-                                "x": 2,
-                                "y": "2",
-                            },
-                        ),
-                        Bbox(
-                            1,
-                            2,
-                            2,
-                            2,
-                            label=3,
-                            attributes={
-                                "score": 0.5,
-                            },
-                        ),
-                        Bbox(
-                            5,
-                            6,
-                            2,
-                            2,
-                            attributes={
-                                "x": 2,
-                                "y": "3",
-                                "occluded": False,
-                            },
-                        ),
-                        Ellipse(
-                            5,
-                            6,
-                            2,
-                            2,
-                            attributes={
-                                "x": 2,
-                                "y": "3",
-                                "occluded": False,
-                            },
-                        ),
-                    ],
-                ),
-                DatasetItem(id=3),
-                DatasetItem(id="2.2", media=Image.from_numpy(data=np.ones((2, 4, 3)))),
-            ],
-            categories=["label_%s" % i for i in range(4)],
-        )
-
-        expected = {
-            "images count": 4,
-            "annotations count": 11,
-            "unannotated images count": 2,
-            "unannotated images": ["3", "2.2"],
-            "annotations by type": {
-                "label": {
-                    "count": 2,
-                },
-                "polygon": {
-                    "count": 0,
-                },
-                "polyline": {
-                    "count": 0,
-                },
-                "bbox": {
-                    "count": 4,
-                },
-                "mask": {
-                    "count": 1,
-                },
-                "points": {
-                    "count": 1,
-                },
-                "caption": {
-                    "count": 2,
-                },
-                "cuboid_3d": {"count": 0},
-                "super_resolution_annotation": {"count": 0},
-                "depth_annotation": {"count": 0},
-                "ellipse": {"count": 1},
-                "hash_key": {"count": 0},
-                "feature_vector": {"count": 0},
-                "tabular": {"count": 0},
-                "unknown": {"count": 0},
-            },
-            "annotations": {
-                "labels": {
-                    "count": 6,
-                    "distribution": {
-                        "label_0": [1, 1 / 6],
-                        "label_1": [0, 0.0],
-                        "label_2": [3, 3 / 6],
-                        "label_3": [2, 2 / 6],
-                    },
-                    "attributes": {
-                        "x": {
-                            "count": 2,  # annotations with no label are skipped
-                            "values count": 2,
-                            "values present": ["1", "2"],
-                            "distribution": {
-                                "1": [1, 1 / 2],
-                                "2": [1, 1 / 2],
-                            },
-                        },
-                        "y": {
-                            "count": 2,  # annotations with no label are skipped
-                            "values count": 1,
-                            "values present": ["2"],
-                            "distribution": {
-                                "2": [2, 2 / 2],
-                            },
-                        },
-                        # must not include "special" attributes like "occluded"
-                    },
-                },
-                "segments": {
-                    "avg. area": (4 * 2 + 9 * 1) / 3,
-                    "area distribution": [
-                        {"min": 4.0, "max": 4.5, "count": 2, "percent": 2 / 3},
-                        {"min": 4.5, "max": 5.0, "count": 0, "percent": 0.0},
-                        {"min": 5.0, "max": 5.5, "count": 0, "percent": 0.0},
-                        {"min": 5.5, "max": 6.0, "count": 0, "percent": 0.0},
-                        {"min": 6.0, "max": 6.5, "count": 0, "percent": 0.0},
-                        {"min": 6.5, "max": 7.0, "count": 0, "percent": 0.0},
-                        {"min": 7.0, "max": 7.5, "count": 0, "percent": 0.0},
-                        {"min": 7.5, "max": 8.0, "count": 0, "percent": 0.0},
-                        {"min": 8.0, "max": 8.5, "count": 0, "percent": 0.0},
-                        {"min": 8.5, "max": 9.0, "count": 1, "percent": 1 / 3},
-                    ],
-                    "pixel distribution": {
-                        "label_0": [0, 0.0],
-                        "label_1": [0, 0.0],
-                        "label_2": [4, 4 / 17],
-                        "label_3": [13, 13 / 17],
-                    },
-                },
-            },
-        }
-
-        actual = compute_ann_statistics(dataset)
-
-        self.assertEqual(expected, actual)
-
-    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_stats_with_empty_dataset(self):
-        dataset = Dataset.from_iterable(
-            [
-                DatasetItem(id=1),
-                DatasetItem(id=3),
-            ],
-            categories=["label_%s" % i for i in range(4)],
-        )
-
-        expected = {
-            "images count": 2,
-            "annotations count": 0,
-            "unannotated images count": 2,
-            "unannotated images": ["1", "3"],
-            "annotations by type": {
-                "label": {
-                    "count": 0,
-                },
-                "polygon": {
-                    "count": 0,
-                },
-                "polyline": {
-                    "count": 0,
-                },
-                "bbox": {
-                    "count": 0,
-                },
-                "mask": {
-                    "count": 0,
-                },
-                "points": {
-                    "count": 0,
-                },
-                "caption": {
-                    "count": 0,
-                },
-                "cuboid_3d": {"count": 0},
-                "super_resolution_annotation": {"count": 0},
-                "depth_annotation": {"count": 0},
-                "ellipse": {"count": 0},
-                "hash_key": {"count": 0},
-                "feature_vector": {"count": 0},
-                "tabular": {"count": 0},
-                "unknown": {"count": 0},
-            },
-            "annotations": {
-                "labels": {
-                    "count": 0,
-                    "distribution": {
-                        "label_0": [0, 0.0],
-                        "label_1": [0, 0.0],
-                        "label_2": [0, 0.0],
-                        "label_3": [0, 0.0],
-                    },
-                    "attributes": {},
-                },
-                "segments": {
-                    "avg. area": 0.0,
-                    "area distribution": [],
-                    "pixel distribution": {
-                        "label_0": [0, 0.0],
-                        "label_1": [0, 0.0],
-                        "label_2": [0, 0.0],
-                        "label_3": [0, 0.0],
-                    },
-                },
-            },
-        }
-
-        actual = compute_ann_statistics(dataset)
-        self.assertEqual(expected, actual)
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_unique_image_count(self):
